+++ lib/tello.ex
@@ -6,14 +6,14 @@ defmodule Tello do
 
   ```elixir
   # Start a client
-  {:ok, tello_client} = Tello.start_client({192, 168, 10, 1}, 8889)
+  {:ok, supervisor, client, status_listener} = Tello.start()
 
   # Enable the Tello's SDK mode
-  :ok = Tello.Command.enable(tello_client)
+  :ok = Tello.Command.enable(client)
 
   # Control Tello
-  :ok = Tello.Command.get_sdk_version(tello_client)
-  :ok = Tello.Command.takeoff(tello_client)
+  :ok = Tello.Command.get_sdk_version(client)
+  :ok = Tello.Command.takeoff(client)
   ```
 
   Please check `Tello.Command` for a full list of commands.
@@ -24,7 +24,7 @@ defmodule Tello do
   flowchart TD
     Application(Tello.Application)
     Command(Tello.Command)
-    Tello2(((Ryze Tello)))
+    RyzeTello(((Ryze Tello)))
 
     Application --- Manager
     Command --- Client1
@@ -33,13 +33,23 @@ defmodule Tello do
     subgraph TelloClient[Client]
       Manager(Tello.Client.Manager)
 
-      Manager -.- Client1
-      Manager -.- Client2
+      Manager -.- Supervisor1
+      Manager -.- Supervisor2
 
       subgraph ClientCluster[Cluster]
+        Supervisor1(Tello.Client.Supervisor)
+        Supervisor2(Tello.Client.Supervisor)
+
         Client1(Tello.Client)
         Client2(Tello.Client)
+
+        StatusListener2(Tello.Client.StatusListener)
+
         More(...)
+
+        Supervisor1 -.- Client1
+        Supervisor2 -.- Client2
+        Supervisor2 -.- StatusListener2
       end
     end
 
@@ -54,13 +64,13 @@ defmodule Tello do
         Processor -.- ControlUnit
       end
 
-      Memory --- Processor
-
-      Gateway --- Processor
+      Memory -.- Processor
+      Gateway -.- Processor
     end
 
     Client1 --- Gateway
-    Client2 --- Tello2
+    Client2 --- RyzeTello
+    StatusListener2 --- RyzeTello
   ```
 
   ## See also
@@ -71,22 +81,40 @@ defmodule Tello do
 
   alias Tello.Client.Manager
 
-  @doc """
-  Start a new `Tello.Client`.
-  """
-  @spec start_client(:inet.ip_address(), :inet.port_number()) ::
-          :ignore | {:error, any} | {:ok, pid} | {:ok, pid, any}
-  def start_client(ip, port) do
-    spec = {Tello.Client, {ip, port}}
+  @default_arg [
+    client: [ip: {192, 168, 10, 1}, port: 8889],
+    status_listener: [port: 8890]
+  ]
 
-    DynamicSupervisor.start_child(Manager, spec)
+  @type init_arg :: [
+          client: [ip: :inet.ip_address(), port: :inet.port_number()],
+          status_listener: [port: :inet.port_number()] | nil
+        ]
+
+  @spec start(init_arg) :: {:ok, supervisor :: pid, client :: pid, status_listener :: nil | pid}
+  @doc """
+  Start a new `Tello.Client.Supervisor`.
+  """
+  def start(init_arg \\ @default_arg) do
+    spec = {Tello.Client.Supervisor, init_arg}
+
+    case DynamicSupervisor.start_child(Manager, spec) do
+      {:ok, supervisor_pid} ->
+        client_pid = Tello.Client.Supervisor.pid_for(supervisor_pid, :client)
+        status_listener_pid = Tello.Client.Supervisor.pid_for(supervisor_pid, :status_listener)
+
+        {:ok, supervisor_pid, client_pid, status_listener_pid}
+
+      other ->
+        other
+    end
   end
 
   @doc """
-  Terminate a `Tello.Client`.
+  Terminate a `Tello.Client.Supervisor` and it's children.
   """
-  @spec terminate_client(pid) :: :ok | {:error, :not_found}
-  def terminate_client(tello_client) do
-    DynamicSupervisor.terminate_child(Manager, tello_client)
+  @spec terminate(pid) :: :ok | {:error, :not_found}
+  def terminate(supervisor_pid) do
+    DynamicSupervisor.terminate_child(Manager, supervisor_pid)
   end
 end