+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## main
 
+- Added `Tello.Client.StatusListener` to receive status update from Tello.
+
 ## V0.1.0
 
 - Initial release